Task almost done

+++ app/Http/Controllers/PostController.php
@@ -3,44 +3,89 @@
 namespace App\Http\Controllers;
 
 use App\Models\Post;
+use App\Models\PostPhotos;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PostController extends Controller
 {
     public function index()
     {
-        return Post::all();
+        return response()->json(
+            Post::with('User', 'Photos', 'Likes')
+                ->withCount('Comments','Likes')
+                ->latest()->paginate(10)
+            ,200);
     }
 
-    public function store(Request $request)
+    public function create(Request $request)
     {
-        $request->validate([
-
-        ]);
-
-        return Post::create($request->validated());
+        $data = new Post();
+        return $this->extracted($request, $data);
     }
 
-    public function show(Post $post)
+    public function show($id)
     {
-        return $post;
+        $data = Post::find($id);
+        return response()->json($data);
     }
 
-    public function update(Request $request, Post $post)
+    public function update(Request $request)
     {
-        $request->validate([
-
-        ]);
-
-        $post->update($request->validated());
-
-        return $post;
+        $data = Post::find($request->id);
+        return $this->extracted($request, $data);
     }
 
-    public function destroy(Post $post)
+    public function delete($id)
     {
-        $post->delete();
+        $post = Post::find($id);
+        if ($post) {
+            $post->Photos()->delete();
+            $post->Likes()->delete();
+            $post->Comments()->delete();
+            $post->delete();
+        }
 
-        return response()->json();
+        return response()->json(['status' => 'success'], 200);
+    }
+
+    /**
+     * @param Request $request
+     * @param Post|array|null $data
+     * @return JsonResponse
+     */
+    public function extracted(Request $request, Post|array|null $data): JsonResponse
+    {
+        $data->user_id = Auth::id();
+        if ($request->post) {
+            $data->post = $request->post;
+        }
+
+        $data->save();
+        if ($request->photos) {
+            $photosData = json_decode($request->photos, true);
+            foreach ($photosData as  $photo) {
+                $postPhotos = new PostPhotos();;
+                $postPhotos->post_id = $data->id;
+                $postPhotos->photo_url = $photo['photo_url'];
+                $postPhotos->save();
+            }
+        }
+        return response()->json(['status' => 'success'], 200);
+    }
+
+    public function postPhotoDelete(Request $request)
+    {
+        if ($request->photos) {
+            $photosData = json_decode($request->photos, true);
+            foreach ($photosData as $key => $data2) {
+                $postPhotos = PostPhotos::find($data2['id']);
+                if ($postPhotos) {
+                    $postPhotos->delete();
+                }
+            }
+        }
+        return response()->json(['status' => 'success'], 200);
     }
 }
